device refactor: lift shared logic for qvm qpu out of ForestDevice into new QuantumComputerDevice
